Address the SonarCloud findings on new code
The quality gate failed on new_reliability_rating alone, and on one Bug:
two int sizes were added and only then widened to the long the counts
record holds. Overflow needs 2^31 certificates, so there was no live
risk, but the cast is free.

The rollback annotation is worth more than the rule that asked for it.
createDiscoveredCertificateAtomic declares three checked exceptions, and
the orchestrator's catch relies on a checked failure leaving its
transaction clean so a shaped reason can still commit -- correct today
only because Spring defaults that way. noRollbackFor pins it, so a
class-level rollbackFor cannot silently reverse the caller's reasoning.

importAndReport took eight parameters; the caller now prepares the groups
and the run context, which brings it to five and puts the preparation
next to the rest of the run setup. The consumption loop leaves by one
exit instead of two: fetching the next result reports absence rather than
breaking, and carries the explanation of why a stream failure stops
consumption rather than propagating.

Tests: assertions chained where they share a subject, and both writer
transaction tests now pin which exception they expected -- the writer call
cannot leave the lambda, since it has to run inside the doomed
transaction, but asserting the message removes the false pass.

Not taken: the self-invocation of a transactional method in
CertificateHandler predates this branch and is inert here (a REQUIRED
callee joining an active REQUIRES_NEW caller is what bypassing the proxy
already does), and a proper fix is a writer-bean split -- filed on #1900.
The handler's own checked-exception rule covers a path whose transaction
is programmatic. OffsetDateTime.now() without a zone is the file's
established pattern in four other places.

+++ src/main/java/com/otilm/core/events/handlers/discovery/DiscoveryRunAccumulator.java
@@ -82,7 +82,7 @@ public class DiscoveryRunAccumulator {
     public DiscoveryRunCounts counts() {
         return new DiscoveryRunCounts(
                 contentIdsWithInventoryGap.size(),
-                keyFailureReasons.size() + contentIdsWithKeyGap.size(),
+                (long) keyFailureReasons.size() + contentIdsWithKeyGap.size(),
                 contentIdsNotAttempted.size(),
                 bookkeepingFailures);
     }
